+++ mock-cli-tests/conftest.py
@@ -1,0 +1,8 @@
+from click.testing import CliRunner
+
+import pytest
+
+
+@pytest.fixture
+def runner():
+    return CliRunner()